C:\Users\zainu\Desktop\курс\proekt2\Ozhaev Zaynudin Sharik4.cpp -o C:\Users\zainu\Desktop\курс\proekt2\Ozhaev Zaynudin Sharik4.exe  -Wshadow -Winit-self -Wredundant-decls -Wcast-align -Wundef -Wfloat-equal -Winline -Wunreachable-code -Wmissing-declarations -Wmissing-include-dirs -Wswitch-enum -Wswitch-default -Weffc++ -Wmain -Wextra -Wall -g -pipe -fexceptions -Wcast-qual -Wconversion -Wctor-dtor-privacy -Wempty-body -Wformat-security -Wformat=2 -Wignored-qualifiers -Wlogical-op -Wmissing-field-initializers -Wnon-virtual-dtor -Woverloaded-virtual -Wpointer-arith -Wsign-promo -Wstack-usage=8192 -Wstrict-aliasing -Wstrict-null-sentinel -Wtype-limits -Wwrite-strings -D_DEBUG -D_EJUDGE_CLIENT_SIDE -D_EJC

+++ Ozhaev Zaynudin Sharik4.cpp
@@ -120,7 +120,7 @@ void MoveBall()
       		int Vt2 = -vx1*e+vy1*s;
 
 
-      		int o = Vn2;
+      		float o = Vn2;
       		Vn2 = Vn1;
       		Vn1 = o;
